+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to Erimil will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unrelased
+
+### Added
+- F key shortcut in thumbnail grid to open Slide Mode directly (skips preview)
+
+
 ## [0.3.1] - 2025-12-27
 
 ### Added